made bigint an optional feature
