+++ Cargo.toml
@@ -5,10 +5,11 @@ authors = ["Martin Billinger <flkazemakase@gmail.com>"]
 edition = "2018"
 
 [features]
-default = ["debugging"]
+default = ["debugging", "bigint"]
 thread-safe = []
 debugging = ["thread-safe", "source-tracking"]
 source-tracking = []
+bigint = ["num-bigint"]
 
 [dependencies]
 gfx = "0.18.1"
@@ -19,7 +20,7 @@ imgui-sys = "0.0.23"
 imgui-gfx-renderer = "0.0.23"
 imgui-winit-support = "0.0.23"
 lazy_static = "1"
-num-bigint = { version = "0.3", features = ["rand"] }
+num-bigint = { optional = true, version = "0.3", features = ["rand"] }
 rand = "0.7"
 rustyline = "5"
